requirements.txt
numpy>=1.21.0
absl-py>=0.9.0
alibi>=0.9.6
astor==0.8.1
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
blis==0.7.8
catalogue>=2.0.4,<2.1.0
certifi==2021.5.30
cffi==1.14.6
changefinder==0.3
chardet==4.0.0
Click==8.0.1
colorama==0.4.4
colorful==0.5.4
cycler==0.10.0
cymem==2.0.5
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
gast==0.4.0
gdown==4.4.0
google-pasta==0.2.0
grpcio==1.36.1
h5py>=3.6.0
idna==3.3
imageio==2.9.0
importlib-metadata==4.0.1
ipykernel
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client
jupyter-console==6.4.0
jupyter-core>=4.9.2
Keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kmeans-pytorch==0.3
liac-arff==2.5.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
metakernel==0.29.0
mistune==0.8.4
mne==0.23.0
mock==4.0.3
more-itertools==8.8.0
nbconvert==6.1.0
nbformat==5.1.3
networkx==2.6.3
nose==1.3.7
notebook==6.4.6
octave-kernel==0.32.0
olefile==0.46
opencv-python==4.5.5.64
openpyxl==3.0.9
pandas
pandocfilters==1.5.0
parso==0.8.2
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plac==1.1.3
plotly==5.1.0
plotly-express==0.4.1
prettyprinter==0.18.0
prometheus-client==0.11.0
prompt-toolkit==3.0.18
protobuf==3.17.3
ptyprocess==0.7.0
pycparser==2.20
pydicom==2.2.0
Pygments==2.10.0
pylibjpeg==1.3.0
pylibjpeg-rle==1.1.0
pyparsing==2.4.7
pyreadline==2.1
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-gdcm==3.0.9.0
pytz==2021.1
PyWavelets==1.1.1
pywinpty==0.5.7
PyYAML
pyzmq==22.3.0
qtconsole==5.1.1
requests==2.26.0
retrying==1.3.3
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
scikit-learn
scipy
seaborn==0.11.2
Send2Trash==1.8.0
shap==0.39.0
six==1.16.0
sklearn==0.0
soupsieve==2.3.1
spacy==3.1.1
spacy-lookups-data==0.3.2
srsly==2.4.2
statsmodels
tensorboard
tensorflow
tensorflow-estimator
termcolor==1.1.0
terminado==0.9.4
testpath==0.5.0
Theano==1.0.4
thinc>=8.0.8
threadpoolctl==2.2.0
tornado==6.1
traitlets==5.1.1
typing-extensions==3.10.0.2
urllib3==1.26.6
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
wfdb==3.1.0
wget==3.2
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.5.0
tifffile==2021.8.8
attrs==21.2.0
GPUtil==1.4.0
